Extract CfdiCreator logic to CfdiCreatorTrait
